fix(host): A1 attach validation — --grove format check, dead field, shared parser
- attachCommand now format-validates --grove via isGroveEraId(groveId,
  'grove'), mirroring the existing --project-id check. A malformed
  grove id previously reached the registry write unchecked.
- Remove DetachOptions.mycoHome: detachProject/resolveAttach take no
  mycoHome param, so the field (and the mycoHome test callers passed
  into detachCommand) did nothing.
- Extract the parseFlags argv parser — byte-identical between
  cli/attach.ts and cli/join.ts — into cli/shared.ts so the two member
  commands share one implementation instead of two hand-mirrored ones.
